// src/app/common/interfaces/Ecategory.ts
export enum ECategory {
  MICRO_URBAINES = 'MICRO_URBAINES',
  MINI_CITADINES = 'MINI_CITADINES',
  CITADINES_POLYVALENTES = 'CITADINES_POLYVALENTES',
  COMPACTES = 'COMPACTES',
  BERLINES_TAILLE_S = 'BERLINES_TAILLE_S',
  BERLINES_TAILLE_M = 'BERLINES_TAILLE_M',
  BERLINES_TAILLE_L = 'BERLINES_TAILLE_L',
  SUV_TT_PICKUP = 'SUV_TT_PICKUP'
}